리팩토링 - 모든 Fragment 는 BaseFragment 를 상속받게 함
보일러 플레이트 코드 제거
스크린 진입 시 자동으로 해당 이름으로 이벤트 전송
