+++ app/src/main/java/com/trueedu/project/ui/view/setting/ScreenSettingFragment.kt
@@ -1,10 +1,5 @@
 package com.trueedu.project.ui.view.setting
 
-import android.app.Dialog
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -18,28 +13,20 @@ import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Switch
 import androidx.compose.material3.SwitchDefaults
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.mutableStateOf
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.ComposeView
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.fragment.app.FragmentManager
-import com.google.android.material.bottomsheet.BottomSheetBehavior
-import com.google.android.material.bottomsheet.BottomSheetDialog
-import com.google.android.material.bottomsheet.BottomSheetDialogFragment
-import com.trueedu.project.R
-import com.trueedu.project.analytics.TrueAnalytics
-import com.trueedu.project.data.ScreenControl
 import com.trueedu.project.repository.local.Local
+import com.trueedu.project.ui.BaseFragment
 import com.trueedu.project.ui.common.BackTitleTopBar
 import com.trueedu.project.ui.common.BasicText
-import com.trueedu.project.ui.theme.TrueProjectTheme
 import dagger.hilt.android.AndroidEntryPoint
 import javax.inject.Inject
 
 @AndroidEntryPoint
-class ScreenSettingFragment: BottomSheetDialogFragment() {
+class ScreenSettingFragment: BaseFragment() {
     companion object {
         private val TAG = ScreenSettingFragment::class.java.simpleName
 
@@ -53,75 +40,38 @@ class ScreenSettingFragment: BottomSheetDialogFragment() {
     }
 
     @Inject
-    lateinit var screen: ScreenControl
-
-    @Inject
     lateinit var local: Local
 
-    @Inject
-    lateinit var trueAnalytics: TrueAnalytics
-
-    val a = mutableStateOf(false)
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setStyle(STYLE_NORMAL, R.style.FillScreenSheetTheme)
-
-        trueAnalytics.enterView("screen_setting__enter")
-    }
-
-    override fun onCreateDialog(savedInstanceState: Bundle?): Dialog {
-        return super.onCreateDialog(savedInstanceState).apply {
-            (this as? BottomSheetDialog)?.behavior?.run {
-                this.skipCollapsed = true
-                this.state = BottomSheetBehavior.STATE_EXPANDED
-                this.isDraggable = true
-            }
-        }
-    }
-
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View {
-        return ComposeView(requireContext()).apply {
-            layoutParams = ViewGroup.LayoutParams(
-                ViewGroup.LayoutParams.MATCH_PARENT,
-                ViewGroup.LayoutParams.MATCH_PARENT
-            )
-            setContent {
-                TrueProjectTheme(
-                    n = screen.theme.intValue,
-                    forceDark = screen.forceDark.value
+    @Composable
+    override fun BodyScreen() {
+        Scaffold(
+            topBar = { BackTitleTopBar("스크린 설정", ::dismissAllowingStateLoss) },
+            modifier = Modifier
+                .fillMaxSize()
+                .background(color = MaterialTheme.colorScheme.background),
+        ) { innerPadding ->
+            Column(
+                modifier = Modifier
+                    .fillMaxSize()
+                    .padding(innerPadding)
+            ) {
+                Row(
+                    horizontalArrangement = Arrangement.SpaceBetween,
+                    verticalAlignment = Alignment.CenterVertically,
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .height(56.dp)
+                        .padding(horizontal = 16.dp)
                 ) {
-                    Scaffold(
-                        topBar = { BackTitleTopBar("스크린 설정", ::dismissAllowingStateLoss) },
-                        modifier = Modifier
-                            .fillMaxSize()
-                            .background(color = MaterialTheme.colorScheme.background),
-                    ) { innerPadding ->
-                        Column(
-                            modifier = Modifier
-                                .fillMaxSize()
-                                .padding(innerPadding)
-                        ) {
-                            Row(
-                                horizontalArrangement = Arrangement.SpaceBetween,
-                                verticalAlignment = Alignment.CenterVertically,
-                                modifier = Modifier
-                                    .fillMaxWidth()
-                                    .height(56.dp)
-                                    .padding(horizontal = 16.dp)
-                            ) {
-                                BasicText("강제 다크모드", fontSize = 16, color = MaterialTheme.colorScheme.primary)
-                                MySwitch(
-                                    checked = screen.forceDark.value,
-                                    onCheckedChange = {
-                                        screen.forceDark.value = it
-                                        local.forceDark = it
-                                        trueAnalytics.clickToggleButton("screen_setting__force_dark__click", !it)
-                                    }
-                                )
-                            }
+                    BasicText("강제 다크모드", fontSize = 16, color = MaterialTheme.colorScheme.primary)
+                    MySwitch(
+                        checked = screen.forceDark.value,
+                        onCheckedChange = {
+                            screen.forceDark.value = it
+                            local.forceDark = it
+                            trueAnalytics.clickToggleButton("screen_setting__force_dark__click", !it)
                         }
-                    }
+                    )
                 }
             }
         }
